+++ src/main/php/lang/ast/emit/PHP.class.php
@@ -237,55 +237,16 @@ abstract class PHP extends Emitter {
   }
 
   protected function emitArray($result, $array) {
-    if (empty($array->values)) {
-      $result->out->write('[]');
-      return;
-    }
-
-    $unpack= false;
+    $result->out->write('[');
     foreach ($array->values as $pair) {
-      if ('unpack' === $pair[1]->kind) {
-        $unpack= true;
-        break;
+      if ($pair[0]) {
+        $this->emitOne($result, $pair[0]);
+        $result->out->write('=>');
       }
+      $this->emitOne($result, $pair[1]);
+      $result->out->write(',');
     }
-
-    if ($unpack) {
-      $result->out->write('array_merge([');
-      foreach ($array->values as $pair) {
-        if ($pair[0]) {
-          $this->emitOne($result, $pair[0]);
-          $result->out->write('=>');
-        }
-        if ('unpack' === $pair[1]->kind) {
-          if ('array' === $pair[1]->expression->kind) {
-            $result->out->write('],');
-            $this->emitOne($result, $pair[1]->expression);
-            $result->out->write(',[');
-          } else {
-            $t= $result->temp();
-            $result->out->write('],('.$t.'=');
-            $this->emitOne($result, $pair[1]->expression);
-            $result->out->write(') instanceof \Traversable ? iterator_to_array('.$t.') : '.$t.',[');
-          }
-        } else {
-          $this->emitOne($result, $pair[1]);
-          $result->out->write(',');
-        }
-      }
-      $result->out->write('])');
-    } else {
-      $result->out->write('[');
-      foreach ($array->values as $pair) {
-        if ($pair[0]) {
-          $this->emitOne($result, $pair[0]);
-          $result->out->write('=>');
-        }
-        $this->emitOne($result, $pair[1]);
-        $result->out->write(',');
-      }
-      $result->out->write(']');
-    }
+    $result->out->write(']');
   }
 
   protected function emitParameter($result, $parameter) {